+++ data/all_weights.csv
@@ -12,8 +12,8 @@ date,weight,bmi
 2019-12-06,104.3,35.63
 2019-12-06,104.1,35.57
 2019-12-07,103.6,35.43
-2019-12-08,103.9,35.5
 2019-12-08,103.5,35.36
+2019-12-08,103.9,35.5
 2019-12-09,103.6,35.43
 2019-12-10,103.3,35.29
 2019-12-10,103.9,35.5
@@ -25,10 +25,10 @@ date,weight,bmi
 2019-12-13,103.6,35.43
 2019-12-13,102.3,34.95
 2019-12-14,102.9,35.16
-2019-12-15,102.9,35.16
-2019-12-15,103.3,35.29
-2019-12-15,103.5,35.36
 2019-12-15,102.3,34.95
+2019-12-15,103.5,35.36
+2019-12-15,103.3,35.29
+2019-12-15,102.9,35.16
 2019-12-16,103.5,35.36
 2019-12-16,102.9,35.16
 2019-12-17,103.3,35.29
@@ -54,9 +54,9 @@ date,weight,bmi
 2020-01-11,98.1,33.51
 2020-01-13,97.9,33.45
 2020-01-13,97.7,33.38
-2020-01-14,98.7,33.72
 2020-01-14,96.9,33.1
 2020-01-14,98.2,33.58
+2020-01-14,98.7,33.72
 2020-01-15,97.7,33.38
 2020-01-16,97.8,33.45
 2020-01-16,98.6,33.72
@@ -69,8 +69,8 @@ date,weight,bmi
 2020-01-19,96.0,32.83
 2020-01-20,96.2,32.9
 2020-01-20,95.4,32.63
-2020-01-21,96.6,33.04
 2020-01-21,96.4,32.97
+2020-01-21,96.6,33.04
 2020-01-22,97.8,33.45
 2020-01-22,98.4,33.65
 2020-01-23,98.4,33.65
@@ -83,10 +83,10 @@ date,weight,bmi
 2020-01-28,95.8,32.76
 2020-01-29,94.6,32.35
 2020-01-29,96.4,32.97
-2020-01-30,94.4,32.28
 2020-01-30,94.6,32.35
-2020-01-31,96.0,32.83
+2020-01-30,94.4,32.28
 2020-01-31,95.6,32.69
+2020-01-31,96.0,32.83
 2020-02-01,95.8,32.76
 2020-02-01,94.8,32.42
 2020-02-01,94.4,32.28
@@ -112,5 +112,5 @@ date,weight,bmi
 2020-02-14,95.4,32.63
 2020-02-15,94.0,32.15
 2020-02-15,94.2,32.22
-2020-02-16,94.2,32.22
 2020-02-16,93.2,31.87
+2020-02-16,94.2,32.22
